fix: Broken Contacts page to Contact component
